requirements.txt
matplotlib==3.8.4
numpy==1.26.1
pandas==2.1.2
Pillow==10.1.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.1
clip
datasets
openai-clip
python-dotenv
huggingface-hub
torch
opencv-python
transformers
fire
supervision
jupyter_bbox_widget
roboflow 
dataclasses-json 
wandb
ipywidgets
sentencepiece